# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Site des Plan\u00e8tes API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/planets": {
      "get": {
        "tags": [
          "planets"
        ],
        "summary": "List Planets",
        "operationId": "list_planets_api_v1_planets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanetOut"
                  },
                  "type": "array",
                  "title": "Response List Planets Api V1 Planets Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "planets"
        ],
        "summary": "Create Planet",
        "operationId": "create_planet_api_v1_planets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/planets/{planet_id}": {
      "get": {
        "tags": [
          "planets"
        ],
        "summary": "Get Planet",
        "operationId": "get_planet_api_v1_planets__planet_id__get",
        "parameters": [
          {
            "name": "planet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Planet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "planets"
        ],
        "summary": "Update Planet",
        "operationId": "update_planet_api_v1_planets__planet_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "planet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Planet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "planets"
        ],
        "summary": "Delete Planet",
        "operationId": "delete_planet_api_v1_planets__planet_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "planet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Planet Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PlanetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "mantra": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mantra"
          },
          "blason_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blason Url"
          },
          "color_hex": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9A-Fa-f]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Hex"
          },
          "is_competing": {
            "type": "boolean",
            "title": "Is Competing",
            "default": true
          },
          "is_default_for_newcomers": {
            "type": "boolean",
            "title": "Is Default For Newcomers",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PlanetCreate"
      },
      "PlanetOut": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "mantra": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mantra"
          },
          "blason_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blason Url"
          },
          "color_hex": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9A-Fa-f]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Hex"
          },
          "is_competing": {
            "type": "boolean",
            "title": "Is Competing",
            "default": true
          },
          "is_default_for_newcomers": {
            "type": "boolean",
            "title": "Is Default For Newcomers",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "season_score": {
            "type": "integer",
            "title": "Season Score",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "PlanetOut"
      },
      "PlanetUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "mantra": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mantra"
          },
          "blason_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blason Url"
          },
          "color_hex": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9A-Fa-f]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Hex"
          },
          "is_competing": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Competing"
          },
          "is_default_for_newcomers": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default For Newcomers"
          }
        },
        "type": "object",
        "title": "PlanetUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}